requirements.txt
easydict==1.9
numpy==1.17.4
pandas==0.25.3
ruamel.yaml==0.15.87
scikit-learn==0.22
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.1
